Refactor - Shared code between service types
