--- data-fetch/index.tsx
@@ -1,3 +1,0 @@
-export { getServicePaths } from "./static-paths";
-export { getServiceProps } from "./static-props";
-export { getQueryParamsFromCalendly } from "./server-side-props";
