--- data-fetch/server-side-props.tsx
@@ -1,7 +1,0 @@
-export const getQueryParamsFromCalendly = ({ query }) => {
-  return {
-    props: {
-      ...query,
-    },
-  };
-};
